tectonic-with-biber: init at 2.17
The `tectonic` package depends on a version of `biber` that is generally
different from the one in the nixpkgs `texlive` bundle. This wrapper
bundles tectonic with a compatible version of biber.

The wrapper is partially recovered from a previous nixpkgs commit:

  https://github.com/NixOS/nixpkgs/commit/5aa8e9f0f90b0c9bbdf7b18ead18704d1622c509

For biber<=2.17 on perl>=5.36.0 a patch is necessary. This is recovered:

  https://github.com/NixOS/nixpkgs/commit/c784cdbf6b98e3e3aefb678b6ee8309cbb8bee15

Also, set @bryango as the maintainer of the wrapper.

Co-authored-by: Mauricio Collares <mauricio@collares.org>
Co-authored-by: Doron Behar <doron.behar@gmail.com>