biber-for-tectonic: init at 2.17
Introduce `biber-for-tectonic`. The `tectonic` package depends on a
version of `biber` that is generally different from the one in the
nixpkgs `texlive` bundle. This package provides an override of biber
suitable for use with tectonic.

For biber<=2.17 on perl>=5.36.0 a patch is necessary.
This is recovered from a previous nixpkgs commit:

  https://github.com/NixOS/nixpkgs/commit/c784cdbf6b98e3e3aefb678b6ee8309cbb8bee15

Also, set @bryango as maintainer.

Co-authored-by: Mauricio Collares <mauricio@collares.org>
Co-authored-by: Doron Behar <doron.behar@gmail.com>